requirements.txt
streamlit
PyMuPDF
python-docx
PyDrive
oauth2client
